Pre-deployment commit

+++ routes/articles.js
@@ -4,8 +4,11 @@ const router = express.Router();
 /* Bring in Articles Model */
 let Article = require('../models/article');
 
+/* Bring in User Model */
+let User = require('../models/user');
+
 /* Route Add Article */
-router.get('/add', (req, res) => {
+router.get('/add', checkAuthenticated, (req, res) => {
 	res.render('add_article', {
 		title: 'Add Article'
 	});
@@ -14,7 +17,7 @@ router.get('/add', (req, res) => {
 /* Route Add Submit Post */
 router.post('/add', (req, res) => {
 	req.checkBody('title', 'Title is required').notEmpty();
-	req.checkBody('author', 'Author is required').notEmpty();
+	// req.checkBody('author', 'Author is required').notEmpty();
 	req.checkBody('body', 'Body is required').notEmpty();
 
 	// Get Errors
@@ -27,7 +30,7 @@ router.post('/add', (req, res) => {
 	} else {
 		let article = new Article();
 		article.title = req.body.title;
-		article.author = req.body.author;
+		article.author = req.user._id;
 		article.body = req.body.body;
 
 		article.save(err => {
@@ -45,15 +48,20 @@ router.post('/add', (req, res) => {
 /* Route Load edit form */
 router.get('/edit/:id', (req, res) => {
 	Article.findById(req.params.id, (err, article) => {
-		res.render('edit_article', {
-			title: 'Edit',
-			article: article
-		});
+		if (!req.user || article.author != req.user._id) {
+			req.flash('danger', 'You are not authorized to edit this article');
+			res.redirect('/');
+		} else {
+			res.render('edit_article', {
+				title: 'Edit',
+				article: article
+			});
+		}
 	});
 });
 
 /* Route Update Submit Post */
-router.post('/edit/:id', (req, res) => {
+router.post('/edit/:id', checkAuthenticated, (req, res) => {
 	let article = {};
 	article.title = req.body.title;
 	article.author = req.body.author;
@@ -74,23 +82,43 @@ router.post('/edit/:id', (req, res) => {
 
 /* Route Delete */
 router.delete('/:id', (req, res) => {
+	if (!req.user._id) {
+		res.status(500).send();
+	}
 	let query = { _id: req.params.id };
-
-	Article.deleteOne(query, err => {
-		if (err) {
-			console.log(err);
+	Article.findById(req.params.id, (err, article) => {
+		if (article.author != req.user._id) {
+			res.status(500).send();
+		} else {
+			Article.deleteOne(query, err => {
+				if (err) {
+					console.log(err);
+				}
+				res.send('Success');
+			});
 		}
-		res.send('Success');
 	});
 });
 
 /* Route GET single article */
 router.get('/:id', (req, res) => {
 	Article.findById(req.params.id, (err, article) => {
-		res.render('article', {
-			article: article
+		User.findById(article.author, (err, user) => {
+			res.render('article', {
+				article: article,
+				author: user.name
+			});
 		});
 	});
 });
 
+/* Access Control */
+function checkAuthenticated(req, res, next) {
+	if (req.isAuthenticated()) {
+		return next();
+	} else {
+		req.flash('danger', 'Login required!');
+		res.redirect('/users/login');
+	}
+}
 module.exports = router;
